refactor out OOP pattern
Activations, transfers, losses, layers and networks all had an OOP
architecture. Replacing all the traits with structs still works, and the
flexibility we lose is not necessary. Also, there is no need for generic
parameters in the layer and network structs, they are simply wrappers
for matrix multiplication, nothing more, nothing less. Their behaviour
is changed through *function pointers* to the correct
activation/transfer function. Same idea with losses and optimizers.

Maybe this is a mistake, so I will be saving the current main branch
state as branch `oop` (main is currently at the same commit as dev).